Get the devices or for a tenant.

Get the devices or for a tenant.

get_devices(num_rows = NULL, parse_json = TRUE)

Arguments

num_rows
The number of records to return.

parse_json
If TRUE, parse the JSON object into a data frame.

Value

A data.frame if parse_json = TRUE, and a character string otherwise.

Details

Get the devices for a tenant.

If num_rows = NULL (default), all devices are returned.

If parse_json is TRUE, the JSON object is parsed using jsonlite::fromJSON before being returned. The data is converted to a single flattened data frame.

If parse_json is FALSE, the JSON object is returned as a JSON string.

References

Cumulocity Inventory API

Examples

# \donttest{
get_devices()

#>                          owner             creationTime            type
#> 1 maulin.gajjar@softwareag.com 2018-07-24T22:12:19.378Z c8y_SensorPhone
#> 2 maulin.gajjar@softwareag.com 2018-05-10T19:12:14.718Z        c8y_Demo
#> 3 maulin.gajjar@softwareag.com 2019-07-19T06:35:00.757Z            <NA>
#> 4     service_device-simulator 2019-09-30T19:58:59.838Z  c8y_MQTTDevice
#>                lastUpdated           name
#> 1 2018-09-05T18:16:13.936Z     Pixel 2 XL
#> 2 2019-08-26T22:39:22.016Z  A demo device
#> 3 2019-07-19T06:35:00.757Z     DemoDevice
#> 4 2019-10-01T04:53:57.447Z Temperature #1
#>                                                       self    id
#> 1  https://mg.cumulocity.com/inventory/managedObjects/1200  1200
#> 2   https://mg.cumulocity.com/inventory/managedObjects/200   200
#> 3 https://mg.cumulocity.com/inventory/managedObjects/51868 51868
#> 4 https://mg.cumulocity.com/inventory/managedObjects/53700 53700
#>                                                              c8y_SupportedOperations
#> 1                                          c8y_Message, c8y_Relay, c8y_Configuration
#> 2 c8y.Message, c8y.Relay, c8y_Restart, c8y_Configuration, c8y_Software, c8y_Firmware
#> 3                                                                               NULL
#> 4                                                     c8y_Restart, c8y_Configuration
#>                                                                                                                                                          c8y_Notes
#> 1                                                                                                                                                             <NA>
#> 2 This shows how a device looks like in Cumulocity. The data is static, so you will not see much of a change and you cannot run any remote control commands on it.
#> 3                                                                                                                                                             <NA>
#> 4                                                                                                                                                             <NA>
#>                                                                            c8y_SupportedMeasurements
#> 1                                                                                               NULL
#> 2                                                          c8y_SignalStrength, c8y_TemperatureSensor
#> 3 c8y_SignalStrengthWifi, c8y_Acceleration, c8y_Barometer, c8y_Gyroscope, c8y_Luxometer, c8y_Compass
#> 4                                                                                               NULL
#>                                                       additionParents.self
#> 1  https://mg.cumulocity.com/inventory/managedObjects/1200/additionParents
#> 2   https://mg.cumulocity.com/inventory/managedObjects/200/additionParents
#> 3 https://mg.cumulocity.com/inventory/managedObjects/51868/additionParents
#> 4 https://mg.cumulocity.com/inventory/managedObjects/53700/additionParents
#>   additionParents.references
#> 1                       NULL
#> 2                       NULL
#> 3                       NULL
#> 4                       NULL
#>                                                       childDevices.self
#> 1  https://mg.cumulocity.com/inventory/managedObjects/1200/childDevices
#> 2   https://mg.cumulocity.com/inventory/managedObjects/200/childDevices
#> 3 https://mg.cumulocity.com/inventory/managedObjects/51868/childDevices
#> 4 https://mg.cumulocity.com/inventory/managedObjects/53700/childDevices
#>   childDevices.references
#> 1                    NULL
#> 2                    NULL
#> 3                    NULL
#> 4                    NULL
#>                                                       childAssets.self
#> 1  https://mg.cumulocity.com/inventory/managedObjects/1200/childAssets
#> 2   https://mg.cumulocity.com/inventory/managedObjects/200/childAssets
#> 3 https://mg.cumulocity.com/inventory/managedObjects/51868/childAssets
#> 4 https://mg.cumulocity.com/inventory/managedObjects/53700/childAssets
#>   childAssets.references
#> 1                   NULL
#> 2                   NULL
#> 3                   NULL
#> 4                   NULL
#>                                                       childAdditions.self
#> 1  https://mg.cumulocity.com/inventory/managedObjects/1200/childAdditions
#> 2   https://mg.cumulocity.com/inventory/managedObjects/200/childAdditions
#> 3 https://mg.cumulocity.com/inventory/managedObjects/51868/childAdditions
#> 4 https://mg.cumulocity.com/inventory/managedObjects/53700/childAdditions
#>   childAdditions.references
#> 1                      NULL
#> 2                      NULL
#> 3                      NULL
#> 4                      NULL
#>                                                       assetParents.self
#> 1  https://mg.cumulocity.com/inventory/managedObjects/1200/assetParents
#> 2   https://mg.cumulocity.com/inventory/managedObjects/200/assetParents
#> 3 https://mg.cumulocity.com/inventory/managedObjects/51868/assetParents
#> 4 https://mg.cumulocity.com/inventory/managedObjects/53700/assetParents
#>   assetParents.references
#> 1                    NULL
#> 2                    NULL
#> 3                    NULL
#> 4                    NULL
#>                                                       deviceParents.self
#> 1  https://mg.cumulocity.com/inventory/managedObjects/1200/deviceParents
#> 2   https://mg.cumulocity.com/inventory/managedObjects/200/deviceParents
#> 3 https://mg.cumulocity.com/inventory/managedObjects/51868/deviceParents
#> 4 https://mg.cumulocity.com/inventory/managedObjects/53700/deviceParents
#>   deviceParents.references c8y_Mobile.mnc c8y_Mobile.currentBand
#> 1                     NULL            480                Unknown
#> 2                     NULL           <NA>                   <NA>
#> 3                     NULL           <NA>                   <NA>
#> 4                     NULL           <NA>                   <NA>
#>   c8y_Mobile.connType c8y_Mobile.currentOperator c8y_Mobile.tac c8y_Mobile.imei
#> 1             Unknown                                     13826 358035086702820
#> 2                <NA>                       <NA>           <NA> 861145013087177
#> 3                <NA>                       <NA>           <NA>            <NA>
#> 4                <NA>                       <NA>           <NA>            <NA>
#>   c8y_Mobile.cellId c8y_Mobile.mcc     c8y_Mobile.iccid
#> 1          13903894            311                 <NA>
#> 2         4904-A496           <NA> 89490200000876620613
#> 3              <NA>           <NA>                 <NA>
#> 4              <NA>           <NA>                 <NA>
#>   c8y_Availability.lastMessage c8y_Availability.status
#> 1     2018-09-05T18:16:14.035Z             UNAVAILABLE
#> 2     2019-04-05T09:24:52.691Z             MAINTENANCE
#> 3                         <NA>                    <NA>
#> 4                         <NA>                    <NA>
#>   c8y_ActiveAlarmsStatus.major c8y_ActiveAlarmsStatus.critical
#> 1                            0                              NA
#> 2                            1                               0
#> 3                           NA                               0
#> 4                           NA                              NA
#>   c8y_RequiredAvailability.responseInterval c8y_Connection.status
#> 1                                         5          DISCONNECTED
#> 2                                       -15           MAINTENANCE
#> 3                                        NA                  <NA>
#> 4                                        NA                  <NA>
#>   c8y_Position.lng c8y_Position.alt c8y_Position.lat c8y_Position.accuracy
#> 1      -117.233149            21.30         32.91942                    NA
#> 2         6.769786           102.47         51.26744                    NA
#> 3               NA               NA               NA                    NA
#> 4        77.000000            88.00         66.00000                    99
#>   c8y_Hardware.serialNumber c8y_Hardware.model c8y_Hardware.revision
#> 1            801KPTM1359769         Pixel 2 XL                taimen
#> 2          0000000017b769d5     Demo model XYZ                  000e
#> 3                      <NA>               <NA>                  <NA>
#> 4                      <NA>               <NA>                  <NA>
#>   c8y_Firmware.name c8y_Firmware.version c8y_Geofence.lng c8y_Geofence.active
#> 1              <NA>                 <NA>               NA                  NA
#> 2      demofirmware         1.20130207-1         8.627466                TRUE
#> 3              <NA>                 <NA>               NA                  NA
#> 4              <NA>                 <NA>               NA                  NA
#>   c8y_Geofence.radius c8y_Geofence.lat c8y_Software.driverSoftware
#> 1                  NA               NA                        <NA>
#> 2                 400         49.89469        driverSoftware-1.0.2
#> 3                  NA               NA                        <NA>
#> 4                  NA               NA                        <NA>
#>                                                     c8y_Configuration.config
#> 1                                                                       <NA>
#> 2 aConfigurationProperty: aConfigurationValue\nanotherProperty: anotherValue
#> 3                                                                       <NA>
#> 4                                                                       <NA>
#>   c8y_MotionTracking.active
#> 1                        NA
#> 2                     FALSE
#> 3                        NA
#> 4                        NA

# }